Fixed issue of multiple log or prm files in your directory.
Before it was searching for log and prm, now it uses the last one
that it finds, which will normally be the most recent one, via
the log1 log2 philosophy. Users can also rename their files so that
the last one in the search string will come up. This is only an issue
for people with multiple files in the folder.