Remove test_load_nc
test_load_nc was written by Sean Lucey to replace load_nc.
The code is based on packages that are not currently imported and
the function has not been tested, whereas the current function load_nc
is tested and known to work within Atlantisom and so I removing the
redundant function. -KFJ