change how bad values are replaced in calc_Z
Now it is within the loop to make sure that it is done by group.
Also if a good value exists later in the time steps that is used
rather than earlier, because if there are many negative values in a row
than the earlier value would be used multiple times.