Update example run to have 73 day output.
The catch files are not available for the 73 day output, which will
need to be checked in to. I am not sure if fishing was turned off
when Bec ran the model with the new 73 output step.
All examples need to be checked to see if they still run with the
new files. Also the .RData file needs to be updated.